OptymalizacjaProjekt/pyproject.toml
[project]
name = "optymalizacjaprojekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.3",
    "fastapi>=0.136.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "ruff>=0.15.12",
    "uvicorn>=0.46.0",
]
